scribble: fix PDF rendering of references to unnumbered sections
Instead of "section" or "part" without a folowing number,
a reference to an unnumbered section renders as the section
name in quotes and then "on page <n>" (to help clarify that
the quoted part is a section name).

More generally, enable LaTeX macros for `secref' to work with the
LaTeX-level label (to expand to a page number, for example) in
the case of document-local references.

Closes PR 13935

original commit: 68d9da17fa9e9b1957ca6426727773cc4f1bb8ca