Add https://blog.racket-lang.org/2012/06/submodules.html as an example for in-source documentation
I couldn't locate gif.rkt in the `file' collection. See a complete working example was helpful to get started. I think it will be as useful for
other Racketeers .(especially beginners like myself).

If gif.rkt is indeed gone due to refactorings, I guess the reference to it should removed.