Automatically link to source code of document.
This works, provided that:
  - the package is on the catalog at pkgs.racket-lang.org
  - the package is hosted on GitHub or GitLab

Restriction 2 could be lifted for other known sites or packages
 hosted as directories.
Restriction 1 would be harder to lift.

This only links to the _top level_ file that defines the overall
document, not the file defining the particular section. Fixing that
would require the Scribble renderer to provide more detailed
information, although it's certainly a possible extension.

Closes #223
Closes #208
Closes racket/racket#874

Related to #76, #228