Added support for latex escapes in bibtex titles and authors
This allows using certain escape sequences, such as \", in titles and authors of
bibtex files.
Titles are wrapped titles with 'exact-chars, so the set of escapes handled is
large.

A different method is used for author which basically convert certain well-known
escapes into Unicode.
This seems necessary to support parsing of author names.