Use key instead of keyCode
In non QWERTY keyboard, pressing S won't focus the search bar.
This PR fixes the problem.

Note that `keyCode` is not supported in IE8 already, so the fact
that `key` is not supported in IE8 too doesn't really matter:
the PR doesn't cause browser compatibility for the features to degrade.