documenting forms: disallow empty box (#287)
* documenting forms: disallow empty box

Forms like `deftogether` and `defproc*` currently allow empty spec,
causing an empty box to be rendered. These malformed outputs are better
prevented at compile-time.

`defform*` already disallows empty spec, but the error is reported
in terms of `defform*/subs`, so this PR adjusts it to directly error
in terms of `defform*`. `defsubform*` still unfortunately errors
in terms of `defform*`, but the fix is too invasive to be worth it.

This PR is tested against all docs in the main distribution.
There's no error.

* document defthing*

* disallow empty content for the history element