Add table of contents toggle for low-width devices
This adds a button to show / hide the TOC column on low-width devices.

It also adjusts various styles to keep the top bar fixed as you scroll so that
the TOC toggle button remains accessible anywhere in the document. Flexbox is
used in the low-width nav bars to improve vertical centering.

Fixes https://github.com/racket/scribble/issues/313