definterface: allow a `#:no-inherit` annotation on superinterfaces
The `#:no-inherit` annotation can avoid an apparent dependeny on
document A when you actually depend on B and B depends on A. Normally,
that's handled ok automatically, but they way `defclass` and
definterface` full information from superinterfaces, the dependecy can
leak through. A better fix would track the dependency as indirect, but
that's a more complex change.