Add kind information to the index
This PR creates new index types that contain the kind information.
This allows e.g. the search page to display or refine the query
based on the information.

Only `form-index-desc`, `procedure-index-desc`, and `thing-index-desc`
are considered, because they are the only ones that currently
allow kind customization. Other indices are already too specific
(e.g. `mixin`, `struct`), so there seems to be no point to allow
customization in such cases. On the other hand, newer forms
that need the kind customization probably should just use
`thing-index-desc(*)`, similar to how the Rhombus documentation
is currently doing.

To maintain backward compatability, the kind information is not added to
any existing structs. Instead, new index types are created at the leaf
level of the index hierarchy.

It's unclear what the contract of the kind information should be.
Currently, it is `string?`. Here are some other possibilities

- `(or/c string? #f)`: this allows the information to be left out.
  However, I think that anything indexed with form, proc, or thing
  should mandate the kind information, because they default to
  "syntax", "procedure", or "value", so they are already non-empty.
  If one wishes to leave the information out, they can use
  `exported-index-desc` directly (this is what the `syntax/parse`
  library does).

- `xexpr?`: this allows richer encoding. In particular,
  `method-index-desc` currently displays information like
  "get-admin (method of editor<%>)" on the search page.
  This "(method of editor<%>)" is ad-hoc inserted and
  "editor%" is typeseted differently than "method of".
  The idea of `xexpr?` is to make it general enough to typeset
  "(method of editor<%>)" under one single framework,
  without special-casing `method-index-desc` and also allow
  this flexible customization for non-methods.

  On the other hand, allowing `xexpr?` could introduce vulnerabilities
  like HTML injection. We could constrain it, but that seems
  like too much work. So I think continuing to
  special-casing `method-index-desc` might be better.

Therefore, I'm sticking with `string?` for now, until we figure out a
better solution. However, it should be noted that further modification
will risk breaking backward compatibility.

Also bump the version, to be used for `#:version` in other `info.rkt`