update suggested URL for `--redirect-main`
Documentation for linking to Racket documentation supplied a URL based
on an older (prior to v6.0) download-site layout. Unfortunately, due
to a limited supply of redirects for the S3-hosted site, we had to
disable support for the old paths a while back. This commit updates
the Scribble documentation's URL to match the more recent organzation.

Closes racket/racket-lang-org#226