feat: support #:auto-value for defparam and defthing
Manually specifying `#:value` could be cumbersome.
An example is `html-empty-tags` from the `xml` collection,
which is a list of more than 10 symbols.
It is also error-prone, and could be out-of-sync from the actual value.

This PR adds a support for `#:auto-value` in `defparam`
and `defthing` to automatically query the value from
the `for-label` binding.

The original idea of this PR is from
https://github.com/racket/racket/pull/4807.

Thanks to @LiberalArtist who recommended an implementation strategy,
and @rocketnia for feedback.