Add `exported-index-desc*`
The `exported-index-desc*` struct adds an `extras` field that is a
composable catch-all for extensions. The set of keys in `extras` is
meant to be a mixture of well-known keys and third-party keys that a
producer and consumer might agree on; the contract on
`exported-index-desc*` checks that well-known keys have values of the
right shape.

The initial well-known keys are `'kind` (corresponding to the text in
the upper right of a documentation box), `'hidden?` (indicates an
entry that looks redundant to a user, so it should be filtered out
when rendering), `'method`, `'constructor?`, and `class-tag`. Those
last 3 cover the information that used to be provided by
`method-index-desc` and `constructor-index-desc`.

For a method, the index text now includes `(method of <class>)` text
that is otherwise added as a special case for `method-index-desc` in
the documentation search panel. The intent here is to put those
decisions more consistently in the document producer, not the index
consumer. Also, `(struct)` is added to struct type entries, because
there's no naming convention that makes that otherwise apparent.

Co-authored-by: Sorawee Porncharoenwase <sorawee.pwase@gmail.com>