add `index-desc`; add `#:extras` argument to `index` and company
Add a generic `index-desc` that is like `exported-index-desc*` in that
it has an `extras` table. Also, add `#:extras` arguments to `index` and
similar so that extras can be supplied at the indexing site.

The remaining non-`exported-index-desc` structs are meant to be
subsumed by `index-desc`. The two other fields of a
`exported-index-desc*` could be converted to values in the `extra`s
table, too, but there are several clients of `exported-index-desc`,
and it's the most common case of indexing. To balance simplicty
against compatibility, the approach here is to end up with three
non-deprecated structure types: `exported-index-desc`,
`exported-index-desc*`, and `index-desc`.

Add `'language-family` and `'sort-order` known keys for extras. The
`'language-family` key is meant to accomodate language family names
like "Rhombus". The`'sort-order` key is meant to allow configuration
of a order for same-keyed index entries (e.g., so that Rhombus
expression-position bindings can consistently before binding-position
bindings when a name is used for both).

Additional keys include `'display-from-libs` for
`exported-index-desc*` to provide exporting module names in the
module's language's syntax, instead of Racket syntax as in the
`from-libs` field.

Those changes allow an individual index entry to specify details, but
in many case, a document-wide or part-wide default is better. Add a
general notion of part contexts, building on the support for part
tags, and use it to propagate `'index-extras` defaults to individual
entries.