add indirection for user-scope searches
When viewing documenation for a user-scoped package, the search box in
the top left was originally meant to find the documentation entry
point using a `PLT_Root...` cookie, which migrated to mean
`PLT_Root...` local storage. The local-storage approach stopped
working years ago, as browsers made different file count as different
domains.

This change enables a layer like `racket-index` to define cause
`user_doc_root` to be defined, which it can do as part of the
`local-redirect` layer that is alreayd in place. And that makes the
top-left search box work again.