replace use of html tt with span of class 'stt'
HTML5 has deprecated the 'tt' tag. This commit replaces it with a span of class 'stt'. Note that the Scribble @tt function has already been fixed to use a span of class 'tt''. These two remaining uses are styling the nbsp separator between the section heading number and the section heading text. One use is in the tocsub link for the section, and the other is in the section heading itself.