doc search: fill in `PLT_Root` query
Whenever a specific entry point into documentation is given, set
`PLT_Root` to point to the first point in the doc-search path, as long
as "index.html" exists there. This fixes two problems: the given entry
point probably doesn't set `PLT_Root` for further searching, and it
may not be in user scope while other documentation is in user scope.